refactor(frontmatter-lint): split validator into its own module to preserve mi (resolves #1333)
